Refactored extra categories render to be contained inside board controller
